refactor(GamePlayer) : 결과를 확인하는 로직을 게임 플레이어가 가지도록 변경
