+++ src/test/java/blackjack/domain/player/DealerTest.java
@@ -48,33 +48,4 @@ public class DealerTest {
 
         assertThat(sut.isReceivable()).isFalse();
     }
-
-
-    @Test
-    @DisplayName("딜러가 버스트 일때 플레이어가 버스트가 아니면 플레이어가 승리한다.")
-    public void Dealer_Player_win_when_dealer_is_burst() {
-        GamePlayer gamePlayer = PlayerFixture.게임_플레이어_생성(List.of(CardValue.EIGHT, CardValue.THREE));
-        Name name = new Name("딜러");
-        Cards cards = CardFixture.카드_목록_생성(List.of(CardValue.EIGHT, CardValue.FOUR));
-        var sut = new Dealer(name, cards);
-        sut.drawCard(new Card(CardValue.JACK, CardSymbol.HEART));
-
-        var result = sut.giveGamePlayerResult(gamePlayer);
-
-        assertThat(result).isEqualTo(ResultStatus.WIN);
-    }
-
-    @Test
-    @DisplayName("플레이어는 버스트이면 무조건 플레이어가 패배한다.")
-    public void Dealer_Player_lose_when_burst() {
-        GamePlayer gamePlayer = PlayerFixture.게임_플레이어_생성(List.of(CardValue.EIGHT, CardValue.FIVE));
-        Name name = new Name("딜러");
-        Cards cards = CardFixture.카드_목록_생성(List.of(CardValue.EIGHT, CardValue.FOUR));
-        var sut = new Dealer(name, cards);
-        gamePlayer.drawCard(new Card(CardValue.JACK, CardSymbol.CLOVER));
-
-        var result = sut.giveGamePlayerResult(gamePlayer);
-
-        assertThat(result).isEqualTo(ResultStatus.LOSE);
-    }
 }
